r0.0080a - Bugfix
If you set the _LOCAL_FILE variable without a trailing backslash "\" the last folder will be stripped when using a line like

FOR %%G IN ("%_LOCAL_FILE%") DO SET "_LOCAL_FILE_PATH=%%~dpG"

that extracts just the path from the _LOCAL_FILE variable. Since we already know we are RECEIVEing file(s) before performing this check, we know the _LOCAL_FILE variable must be a folder path anyway, the answer is to not convert _LOCAL_FILE into _LOCAL_FILE_PATH at all, just use the variable as-is and treat it as a folder path.